+++ src/som/primitives/actors/CreateActorPrim.java
@@ -1,7 +1,6 @@
 package som.primitives.actors;
 
 import com.oracle.truffle.api.dsl.GenerateNodeFactory;
-import com.oracle.truffle.api.dsl.NodeChild;
 import com.oracle.truffle.api.dsl.Specialization;
 import com.oracle.truffle.api.source.SourceSection;
 
@@ -9,24 +8,29 @@ import som.interpreter.actors.Actor;
 import som.interpreter.actors.SFarReference;
 import som.interpreter.nodes.nary.BinaryComplexOperation;
 import som.primitives.ObjectPrims.IsValue;
-import som.primitives.ObjectPrimsFactory.IsValueFactory;
+import som.primitives.ObjectPrimsFactory.IsValueFactory.IsValueNodeGen;
 import som.primitives.Primitive;
 import som.primitives.actors.PromisePrims.IsActorModule;
 import som.vm.VmSettings;
+import som.vm.constants.KernelObj;
 import tools.concurrency.ActorExecutionTrace;
 
 
 @GenerateNodeFactory
 @Primitive(primitive = "actors:createFromValue:", selector  = "createActorFromValue:",
-           specializer = IsActorModule.class, extraChild = IsValueFactory.class)
-@NodeChild(value = "isValue", type = IsValue.class, executeWith = "receiver")
+           specializer = IsActorModule.class)
 public abstract class CreateActorPrim extends BinaryComplexOperation {
-  protected CreateActorPrim(final boolean eagWrap, final SourceSection source) { super(eagWrap, source); }
+  @Child protected IsValue isValue;
 
-  @Specialization(guards = "isValue")
-  public final SFarReference createActor(final Object nil, final Object value, final boolean isValue) {
+  protected CreateActorPrim(final boolean eagWrap, final SourceSection source) {
+    super(eagWrap, source);
+    isValue = IsValueNodeGen.createSubNode();
+  }
+
+  @Specialization(guards = "isValue.executeEvaluated(argument)")
+  public final SFarReference createActor(final Object receiver, final Object argument) {
     Actor actor = Actor.createActor();
-    SFarReference ref = new SFarReference(actor, value);
+    SFarReference ref = new SFarReference(actor, argument);
 
     if (VmSettings.ACTOR_TRACING) {
       ActorExecutionTrace.actorCreation(ref);
@@ -34,5 +38,8 @@ public abstract class CreateActorPrim extends BinaryComplexOperation {
     return ref;
   }
 
-  // TODO: add a proper error or something if it isn't a value...
+  @Specialization(guards = "!isValue.executeEvaluated(argument)")
+  public final Object throwNotAValueException(final Object receiver, final Object argument) {
+    return KernelObj.signalException("signalNotAValueWith:", argument);
+  }
 }